Issue #OS-158 Moved validation files to io.opensaber.registry.config.validation package
