+++ java/registry/src/main/java/io/opensaber/registry/model/DBConnectionInfo.java
@@ -1,6 +1,6 @@
 package io.opensaber.registry.model;
 
-import io.opensaber.registry.sink.shard.ValidConnectionInfo;
+import io.opensaber.registry.config.validation.ValidConnectionInfo;
 @ValidConnectionInfo
 public class DBConnectionInfo {
